[api] extract JSApi from Api
